refactor: refactor Domain Category
