+++ src/Category/errors/entityValidationErrors.ts
