+++ tsconfig.json
@@ -9,7 +9,7 @@
     "target": "ES2021",
     "sourceMap": true,
     "outDir": "./dist",
-    "baseUrl": "./",
+    "baseUrl": "./src",
     // "paths": {
     //   "@/*": ["./*" ]
     //  },
